lambda/requirements.txt
firebolt-sdk>=1.0.0
pyarrow>=10.0.0

